requirements-test.txt
pytest~=6.0
pytest-aiohttp~=0.3.0
aioresponses~=0.6.0
pytest-cov~=2.0
